moved inspect_string_custom_hotword into separate file
